게시글 수정 삭제 기능 관련 임시 커밋
- 자유 게시판 수정과 삭제 동작을 추가합니다.
- 삭제 동작의 경우 테스트가 끝나지 않은 임시 커밋입니다.
- DB변경사항이 미적용되었습니다. 따라서 이 커밋의 정상 실행을 보장하지 않습니다.

+++ src/main/java/kr/pe/hayarobys/nox/common/tempsave/TempSaveServiceImpl.java
@@ -33,6 +33,11 @@ public class TempSaveServiceImpl implements TempSaveService{
 	}
 	
 	@Override
+	public Integer selectFileGroupNoFromTempSaveByTempSaveNo(Integer tempSaveNo){
+		return tempSaveDAO.selectFileGroupNoFromTempSaveByTempSaveNo(tempSaveNo);
+	}
+	
+	@Override
 	public void insertTempSave(TempSaveVO tempSaveVO){
 		// TODO: 정합성 검사
 		tempSaveDAO.insertTempSave(tempSaveVO);
@@ -45,18 +50,21 @@ public class TempSaveServiceImpl implements TempSaveService{
 		// 원본 임시 저장 정보 조회
 		TempSaveVO tempSaveVO = selectLastTempSaveFromMemNoAndCategoryAndUse(memNo, tempSaveCategory, tempSaveUse);
 		
+		if(tempSaveVO == null){ return; }
+		
+		// DB상에서 임시 저장 레코드 제거
+		deleteTempSaveByTempSaveNo(tempSaveVO.getTempSaveNo());
+		
 		// 물리 경로에서 임시 파일만 제거, DB에서 특정 파일 그룹의 파일 전부 제거, DB에서 파일 그룹 제거
 		uploadService.deleteFileByFileGroupNo(tempSaveVO.getFileGroupNo(), true);
-		
-		// DB상에서 임시 저장 제거
-		deleteTempSaveByTempSaveNo(tempSaveVO.getTempSaveNo());
 	}
-
+		
 	@Override
 	public void deleteTempSaveByTempSaveNo(Integer tempSaveNo){
 		// TODO: 정합성 검사
 		tempSaveDAO.deleteTempSaveByTempSaveNo(tempSaveNo);
 	}
 
+	
 
 }
